chore: moved remove-i18n script to the scripts directory
