+++ package.json
@@ -1,27 +1,28 @@
 {
-	"name": "advanced-astro-i18n",
-	"type": "module",
-	"version": "3.0.2",
-	"scripts": {
-		"dev": "astro dev",
-		"start": "astro dev",
-		"build": "astro build",
-		"preview": "astro preview",
-		"astro": "astro",
-		"remove-demo": "node scripts/remove-demo.js",
-		"remove-dark-mode": "node scripts/remove-dark-mode.js",
-		"remove-decap": "node scripts/remove-decap.js",
-		"config-i18n": "node scripts/config-i18n.js",
-		"create-page": "node scripts/create-page.js"
-	},
-	"dependencies": {
-		"@astrojs/sitemap": "^3.7.2",
-		"astro": "^6.3.3",
-		"astro-icon": "^1.1.5",
-		"autoprefixer": "^10.5.0",
-		"less": "^4.6.4"
-	},
-	"devDependencies": {
-		"slugify": "^1.6.9"
-	}
+  "name": "advanced-astro-i18n",
+  "type": "module",
+  "version": "3.0.2",
+  "scripts": {
+    "dev": "astro dev",
+    "start": "astro dev",
+    "build": "astro build",
+    "preview": "astro preview",
+    "astro": "astro",
+    "remove-demo": "node scripts/remove-demo.js",
+    "remove-dark-mode": "node scripts/remove-dark-mode.js",
+    "remove-decap": "node scripts/remove-decap.js",
+    "config-i18n": "node scripts/config-i18n.js",
+    "create-page": "node scripts/create-page.js",
+    "remove-i18n": "node scripts/remove-i18n.js"
+  },
+  "dependencies": {
+    "@astrojs/sitemap": "^3.7.2",
+    "astro": "^6.3.3",
+    "astro-icon": "^1.1.5",
+    "autoprefixer": "^10.5.0",
+    "less": "^4.6.4"
+  },
+  "devDependencies": {
+    "slugify": "^1.6.9"
+  }
 }
